bulk commit

+++ backo/item.py
@@ -1,6 +1,7 @@
 """
 Module providing the Item() Class
 """
+
 # pylint: disable=wrong-import-position, no-member, import-error, protected-access, wrong-import-order, attribute-defined-outside-init
 
 import sys
@@ -20,8 +21,7 @@ from .status import StatusType
 log = log_system.get_or_create_logger("Item")
 
 sys.path.insert(1, "../../stricto")
-from stricto import Dict, Int, String
-
+from stricto import Dict, String
 
 
 class Item(Dict):  # pylint: disable=too-many-instance-attributes
@@ -34,7 +34,9 @@ class Item(Dict):  # pylint: disable=too-many-instance-attributes
         available arguments
         """
         self.db_handler = None
-        self.meta_data_handler = kwargs.pop('meta_data_handler', StandardMetaDataHandler())
+        self.meta_data_handler = kwargs.pop(
+            "meta_data_handler", StandardMetaDataHandler()
+        )
         self._loaded_object = None
         self._status = StatusType.UNSET
         self._collection = None
@@ -43,17 +45,17 @@ class Item(Dict):  # pylint: disable=too-many-instance-attributes
         # adding _id to the model
         self.add_to_model("_id", String())
 
-        # Setting meta schema
+        # Setting meta schema
         if self.meta_data_handler:
             self.meta_data_handler.append_schema(self)
 
-    def set_db_handler(self, db_connector: DBConnector ):
+    def set_db_handler(self, db_connector: DBConnector) -> None:
         """
         Set or modify the Database Handler
         """
         self.__dict__["db_handler"] = db_connector
 
-    def compute_hash_schema( self, schema: dict ):
+    def compute_hash_schema(self, schema: dict):
         """
         Compute the Hash of the schema
         (to add in meta data to help, on load, to see if the db structure change)
@@ -75,7 +77,7 @@ class Item(Dict):  # pylint: disable=too-many-instance-attributes
         result.__dict__["_locked"] = True
         return result
 
-    def load(self, _id: str, **kwargs):
+    def load(self, _id: str, **kwargs) -> None:
         """
         Read in the database by Id and fill the Data
 
@@ -105,7 +107,7 @@ class Item(Dict):  # pylint: disable=too-many-instance-attributes
 
         # print(f"Load {int(datetime.timestamp(datetime.now()))}", self)
 
-    def reload(self, **kwargs):
+    def reload(self, **kwargs) -> None:
         """
         Reload from DB the object
 
@@ -131,7 +133,7 @@ class Item(Dict):  # pylint: disable=too-many-instance-attributes
 
         self.trigg("loaded", id(self), **kwargs)
 
-    def save(self, **kwargs):
+    def save(self, **kwargs) -> None:
         """
         save the object.
 
@@ -155,8 +157,15 @@ class Item(Dict):  # pylint: disable=too-many-instance-attributes
             kwargs.get("transaction_id"),
         )
 
+        # Check if right to create
+        if self._collection.has_right("modify", self) is not True:
+            raise Error(
+                ErrorType.RIGHT,
+                f"No rights to modify element in collection {self.self._collection}.",
+            )
+
         if self.meta_data_handler:
-            self.meta_data_handler.set_on_save( self )
+            self.meta_data_handler.set_on_save(self)
 
         # Load the previous value in the DB (for transactions and comparison of values )
         if self.__dict__["_loaded_object"] is None:
@@ -169,7 +178,7 @@ class Item(Dict):  # pylint: disable=too-many-instance-attributes
         self.trigg("before_save", id(self), **kwargs)
 
         # print(f"Save {int(datetime.timestamp(datetime.now()))}", self)
-        dict_to_save = self.get_view('save').get_value()
+        dict_to_save = self.get_view("save").get_value()
 
         self.db_handler.save(self._id.get_value(), dict_to_save)
 
@@ -194,7 +203,7 @@ class Item(Dict):  # pylint: disable=too-many-instance-attributes
 
         self.trigg("saved", id(self), **kwargs)
 
-    def delete(self, **kwargs):
+    def delete(self, **kwargs) -> None:
         """
         delete the object in the database
 
@@ -219,6 +228,13 @@ class Item(Dict):  # pylint: disable=too-many-instance-attributes
             kwargs.get("transaction_id"),
         )
 
+        # Check if right to create
+        if self._collection.has_right("delete", self) is not True:
+            raise Error(
+                ErrorType.RIGHT,
+                f"No rights to delete in collection {self.self._collection}.",
+            )
+
         # Send delete event before deletion to do  some stufs
         self.trigg("deletion", id(self), **kwargs)
         self.db_handler.delete_by_id(self._id.get_value())
@@ -242,7 +258,7 @@ class Item(Dict):  # pylint: disable=too-many-instance-attributes
             self.get_value(),
         )
 
-    def create_uniq_id(self):
+    def create_uniq_id(self) -> str:
         """
         Create an _id before creation.
         Depends on the db_connector used. some of them needs _ids
@@ -266,10 +282,17 @@ class Item(Dict):  # pylint: disable=too-many-instance-attributes
             obj,
         )
 
+        # Check if right to create
+        if self._collection.has_right("create") is not True:
+            raise Error(
+                ErrorType.RIGHT,
+                f"No rights to create in collection {self.self._collection}.",
+            )
+
         self.set(obj)
         # Set _meta
         if self.meta_data_handler:
-            self.meta_data_handler.set_on_create( self )
+            self.meta_data_handler.set_on_create(self)
 
         # Set the _id
         self._id = self.create_uniq_id()
